requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
loguru==0.7.2
psutil==5.9.8 